pyproject.toml
[project]
name = "channel-commerce-platform"
version = "0.1.0"
description = "Multi-tenant private channel commerce platform backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.0,<2.0.0",
    "fastapi>=0.115.0,<1.0.0",
    "httpx>=0.28.0,<1.0.0",
    "psycopg[binary]>=3.2.0,<4.0.0",
    "pydantic-settings>=2.7.0,<3.0.0",
    "redis>=5.2.0,<6.0.0",
    "sqlalchemy>=2.0.36,<3.0.0",
    "uvicorn[standard]>=0.32.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0,<9.0.0",
    "pytest-asyncio>=0.24.0,<1.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
